2018.02.27 1. improvements on save and load 2. add a 1cm-scale on canvas 3. frist step of "exporting .ai file" (stiil need to work on the REAL SIZE accuracy)

+++ unit.h
@@ -23,32 +23,19 @@ public:
 
     // item state
     bool Pressed;
-    int xi;
-    int yi;
-    int length = 1;
-    int width = 1;
+//    int xi;
+//    int yi;
+//    int length = 1;
+//    int width = 1;
+    float xi;
+    float yi;
+    float length = 1;
+    float width = 1;
     int de_type = 0;
     int de_xnum = 0;
     int de_ynum = 0;
     QString type;
     QColor color = Qt::white;
-
-//    bool unit_deletemode = false;
-//    bool unit_detailmode = false;
-
-//    //setting parameter
-//    int unit_chip_length_cm = 3;
-//    int unit_chip_width_cm = 2;
-//    int unit_chip_border_mm = 3;
-//    int unit_cp_length_mm = 2;
-//    int unit_cp_width_mm = 4;
-//    int unit_de1_length_mm = 1;
-//    int unit_de1_width_mm = 5;
-//    int unit_de2_length_mm = 3;
-//    int unit_de_spacing_um = 500;
-//    int unit_cp_spacing_um = 300;
-//    int unit_dispenser_mm = 10;
-//    int unit_pix_per_brick;
 
     //units color
     QColor unit_de1_color = QColor(255, 0, 0, 127);
